refactor: del EdgeComputingVisualization i mindre filer
Trekker ut types/config/particles og deler JSX i NetworkControls + NetworkMap.
Hovedkomponenten er nå en orkestrator (513 → 110 linjer). Partiklene regnes ut
med useMemo i stedet for setState-i-effect.
